# tree: fix tuple encoding performance regression